LayoutTests/svg/animations/animate-color-fill-currentColor-expected.txt

SVG 1.1 dynamic animation tests

Tests animation on 'currentColor'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Initial condition:
PASS document.defaultView.getComputedStyle(rect).getPropertyValue('fill') is "#ff0000"
PASS (fillPaint = document.defaultView.getComputedStyle(rect).getPropertyCSSValue('fill')).toString() is "[object SVGPaint]"
PASS fillPaint.paintType is SVGPaint.SVG_PAINTTYPE_CURRENTCOLOR
PASS fillPaint.uri is ""
PASS fillPaint.colorType is SVGColor.SVG_COLORTYPE_CURRENTCOLOR
PASS (fillColor = fillPaint.rgbColor).toString() is "[object RGBColor]"
PASS fillColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 255
PASS fillColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS fillColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0

Half-time condition:
PASS document.defaultView.getComputedStyle(rect).getPropertyValue('fill') is "#804000"
PASS (fillPaint = document.defaultView.getComputedStyle(rect).getPropertyCSSValue('fill')).toString() is "[object SVGPaint]"
PASS fillPaint.paintType is SVGPaint.SVG_PAINTTYPE_CURRENTCOLOR
PASS fillPaint.uri is ""
PASS fillPaint.colorType is SVGColor.SVG_COLORTYPE_CURRENTCOLOR
PASS (fillColor = fillPaint.rgbColor).toString() is "[object RGBColor]"
PASS fillColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 128
PASS fillColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 64
PASS fillColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0

End condition:
PASS document.defaultView.getComputedStyle(rect).getPropertyValue('fill') is "#008000"
PASS (fillPaint = document.defaultView.getComputedStyle(rect).getPropertyCSSValue('fill')).toString() is "[object SVGPaint]"
PASS fillPaint.paintType is SVGPaint.SVG_PAINTTYPE_CURRENTCOLOR
PASS fillPaint.uri is ""
PASS fillPaint.colorType is SVGColor.SVG_COLORTYPE_CURRENTCOLOR
PASS (fillColor = fillPaint.rgbColor).toString() is "[object RGBColor]"
PASS fillColor.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS fillColor.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 128
PASS fillColor.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS successfullyParsed is true

TEST COMPLETE
